requirements.txt
streamlit
PyPDF2
faiss-cpu
sentence-transformers
transformers
torch
openai
langchain
fpdf
python-docx
